+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ## 3.1.0 - 2021-04-27
 
+### Enhancement
 - Refactored Store URLs to utilize craft native functions and work with changed actionTrigger changes (#48)
+
+### Bugfix
+- Fixed an issue with HPP if LogoAsset or BackgroundAsset is null
 
 ## 3.0.2 - 2021-04-27